# packages/shared-ts/openapi/capability-registry.json
{
  "components": {
    "schemas": {
...
      "CapabilityUpsertRequest": {
        "additionalProperties": false,
        "properties": {
          "description_en": {
            "minLength": 1,
            "title": "Description En",
            "type": "string"
          },
          "description_zh": {
            "minLength": 1,
            "title": "Description Zh",
            "type": "string"
          },
          "examples": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Examples",
            "type": "array"
          },
          "k_algo": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "K Algo"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "model_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Model Version",
            "type": "string"
          },
          "provider_id": {
            "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
            "title": "Provider Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "v1",
              "v1_late",
              "v2",
              "audited",
              "shadow"
            ],
            "title": "Status",
            "type": "string"
          },
          "supported_solvers": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Supported Solvers",
            "type": "array"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "task_type": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Task Type",
            "type": "string"
          },
          "tenant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "tier": {
            "pattern": "^(T[1-6]|P[1-5])$",
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "task_type",
          "tier",
          "status",
          "provider_id",
          "model_version",
          "supported_solvers",
          "description_zh",
          "description_en"
        ],
        "title": "CapabilityUpsertRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Story 7.A.1 — strict-minimal provider/capability schema reservation. Does not execute provider marketplace, OAuth, revenue-share, or auto-migration flows.",
    "title": "OptiCloud Capability Registry",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "operationId": "list_capabilities_v1_capabilities_get",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CapabilityResponse"
                  },
                  "title": "Response List Capabilities V1 Capabilities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Capabilities",
        "tags": [
          "capabilities"
        ]
      }
    },
    "/v1/capabilities/{k_algo}": {
      "get": {
        "operationId": "get_capability_v1_capabilities__k_algo__get",
        "parameters": [
          {
            "in": "path",
            "name": "k_algo",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "K Algo",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Capability",
        "tags": [
          "capabilities"
        ]
      },
      "put": {
        "operationId": "upsert_capability_v1_capabilities__k_algo__put",
        "parameters": [
          {
            "in": "path",
            "name": "k_algo",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "K Algo",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Internal-Service-Auth",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Service-Auth"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapabilityUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Capability",
        "tags": [
          "capabilities"
        ]
      }
    },
    "/v1/providers": {
      "get": {
        "operationId": "list_providers_v1_providers_get",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderResponse"
                  },
                  "title": "Response List Providers V1 Providers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Providers",
        "tags": [
          "providers"
        ]
      }
    },
    "/v1/providers/{provider_id}": {
      "get": {
        "operationId": "get_provider_v1_providers__provider_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Provider",
        "tags": [
          "providers"
        ]
      },
      "put": {
        "operationId": "upsert_provider_v1_providers__provider_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Internal-Service-Auth",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Service-Auth"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Provider",
        "tags": [
          "providers"
        ]
      }
    },
...
    "/v1/providers/{provider_id}/oauth-flow/execute": {
      "post": {
        "operationId": "execute_oauth_flow_v1_providers__provider_id__oauth_flow_execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Oauth Flow",
        "tags": [
          "provider-oauth"
        ]
      }
    }
  }
}